requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
prometheus-client==0.21.0
python-dotenv==1.0.1
pydantic==2.9.2
pillow==11.0.0
